nfc-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pyscard==2.0.7
python-dotenv==1.0.0
colorama==0.4.6
requests==2.32.4
